% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureVMmetadata.R
\name{in_azure_vm}
\alias{in_azure_vm}
\alias{get_vm_cert}
\title{Check if R is running in an Azure VM}
\usage{
in_azure_vm()

get_vm_cert(nonce = NULL)
}
\arguments{
\item{nonce}{An optional string to use as a nonce.}
}
\value{
For \code{in_azure_vm}, a boolean. For \code{get_vm_cert}, a PKCS-7 certificate object.
}
\description{
Check if R is running in an Azure VM
}
\details{
These functions check if R is running in an Azure VM by attempting to contact the instance metadata host. \code{in_azure_vm} simply returns TRUE or FALSE based on whether it succeeds. \code{get_vm_cert} provides a stronger check, by retrieving the VM's certificate and throwing an error if this is not found. Note that you should still verify the certificate's authenticity before relying on it.
}
