% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureVMmetadata.R
\docType{data}
\name{instance}
\alias{instance}
\alias{attested}
\alias{events}
\alias{update_instance_metadata}
\alias{update_attested_metadata}
\alias{update_scheduled_events}
\title{Metadata for an Azure VM}
\format{An object of class \code{environment} of length 0.}
\usage{
instance

attested

events

update_instance_metadata()

update_attested_metadata(nonce = NULL)

update_scheduled_events()
}
\arguments{
\item{nonce}{For \code{update_attested_metadata}, an optional string to use as a nonce.}
}
\value{
The updating functions return lists of metadata items, invisibly.
}
\description{
Metadata for an Azure VM
}
\details{
The \code{instance}, \code{attested} and \code{events} objects are environments containing the instance metadata, attested document, and scheduled events respectively for a VM running in Azure. \code{instance} and \code{attested} are automatically populated when you load the AzureVMmetadata package, or you can manually populate them yourself with the \code{update_instance_metadata} and \code{update_attested_metadata} functions. \code{events} is not populated at package startup, because calling the scheduled event service can require up to several minutes if it is not running already. You can manually populate it with the \code{update_scheduled_events} function.
}
\seealso{
\href{https://docs.microsoft.com/en-us/azure/virtual-machines/windows/instance-metadata-service}{Instance metadata service documentation}

To obtain OAuth tokens from the metadata service, see \link[AzureAuth:get_managed_token]{AzureAuth::get_managed_token}
}
\keyword{datasets}
